Refactor shared utilities and fix null checks
- Extract `resolveCssVar` to `tools/colors` and re-export from widget-flood-risk/utils
- Update `useClientOnly` import path from `tools/hooks/` to `hooks/` in Heatmap component and tests
- Replace `||` with `??` for null/undefined checks in pressure display (WidgetSummary, WidgetForecastTable)
